+++ requests/models.py
@@ -37,7 +37,7 @@ class Request(object):
     def __init__(self,
         url=None, headers=dict(), files=None, method=None, data=dict(),
         params=dict(), auth=None, cookiejar=None, timeout=None, redirect=False,
-        allow_redirects=False, proxies=None):
+        allow_redirects=False, proxies=None, hooks=None):
 
         #: Float describ the timeout of the request.
         #  (Use socket.setdefaulttimeout() as fallback)
@@ -94,6 +94,8 @@ class Request(object):
         #: True if Request has been sent.
         self.sent = False
 
+        #: Event-handling hooks.
+        self.hooks = hooks
 
         # Header manipulation and defaults.
 
@@ -360,6 +362,8 @@ class Request(object):
             except (urllib2.HTTPError, urllib2.URLError), why:
                 if hasattr(why, 'reason'):
                     if isinstance(why.reason, socket.timeout):
+                        why = Timeout(why)
+                    elif isinstance(why.reason, socket.error):
                         why = Timeout(why)
 
                 self._build_response(why, is_error=True)